chore: rename _octextinputpassword to to octextinputpassword
