+++ packages/design-system/src/components/OcTextInput/OcTextInput.vue
@@ -83,7 +83,7 @@ import { HTMLAttributes, nextTick, computed, useAttrs, useTemplateRef } from 'vu
 import { uniqueId } from '../../helpers'
 import OcButton from '../OcButton/OcButton.vue'
 import OcIcon from '../OcIcon/OcIcon.vue'
-import OcTextInputPassword from '../_OcTextInputPassword/_OcTextInputPassword.vue'
+import OcTextInputPassword from '../OcTextInputPassword/OcTextInputPassword.vue'
 import { PasswordPolicy } from '../../helpers'
 import { useGettext } from 'vue3-gettext'
