+++ packages/design-system/src/components/OcTextInputPassword/OcTextInputPassword.vue
